requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==18.2.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
edge-tts==7.2.3
exceptiongroup==1.3.1
fastapi==0.122.0
filetype==1.2.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
google-genai==1.47.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.80
langchain-google-genai==2.1.12
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
moviepy==2.2.1
multidict==6.7.0
numpy==2.0.2
openai==2.8.1
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pdflatex==0.1.3
pillow==11.3.0
proglog==0.1.12
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.26.5
pyparsing==3.2.5
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
replicate==1.0.7
reportlab==4.4.5
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
starlette==0.49.3
tabulate==0.9.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
sqlalchemy==2.0.36
